refactor : Hit를 Running으로 추상화
